[Feature #163795096] restructured the projects
