+++ blockchain.js
@@ -1,0 +1,104 @@
+const SHA256 = require('crypto-js/sha256');
+
+class Transaction{
+    constructor(fromAddress, toAddress, amount){
+        this.fromAddress = fromAddress;
+        this.toAddress = toAddress;
+        this.amount = amount; 
+    }
+}
+
+class Block {
+    constructor(timestamp,transactions,prevHash = ''){
+        this.timestamp = timestamp;
+        this.transactions = transactions;
+        this.hash = this.calculateHash();
+        this.nonce = 0;
+
+    }
+
+    calculateHash(){
+
+        return SHA256(this.index + this.prevHash + this.timestamp + JSON.stringify(this.transactions) + this.nonce).toString()
+    }
+
+    mineBlock(difficulty){
+        while(this.hash.substring(0, difficulty) !== Array(difficulty + 1).join("0")){
+            this.nonce++;
+            this.hash = this.calculateHash();
+        }
+        console.log("Block mined: " + this.hash)
+    }
+}
+
+
+class Blockchain{
+    constructor(){
+        this.chain = [this.createGenesisBlock()];
+        this.difficulty = 2;
+        this.pendingTransactions = [];
+        this.miningReward = 100;
+    }
+
+    createGenesisBlock(){
+        return new Block("2019-02-03","Genesis block","0");
+    }
+
+    getLatestBlock(){
+        return this.chain[this.chain.length -1];
+    }
+
+    minePendingTransactions(miningRewardAddress){
+        let block = new Block(Date.now(), this.pendingTransactions);
+        block.mineBlock(this.difficulty);
+
+        console.log("Block succesfully mined!");
+        this.chain.push(block);
+
+        this.pendingTransactions = [
+            new Transaction(null, miningRewardAddress, this.miningReward)
+        ];
+    }
+
+    createTransactions(transaction){
+        this.pendingTransactions.push(transaction)
+    }
+
+    getBalancedOfAddress(address){
+        let balance = 0;
+
+        for(const block of this.chain){
+            for(const trans of block.transactions){
+                if(trans.fromAddress === address){
+                    balance -= trans.amount;
+                }
+                if(trans.toAddress === address){
+                    balance += trans.amount;
+                }
+            }
+        }
+
+        return balance;
+    }
+    isChainValid(){
+        for(let i = 1; i < this.chain.length; i++){
+            const currentBlock = this.chain[i];
+            const prevHash = this.chain[i -1];
+
+            if(currentBlock.hash != currentBlock.calculateHash()){
+                return false;
+            }
+
+            if(currentBlock.prevHash != prevHash.hash){
+                return fasle;
+            }
+    
+            return true;
+        }
+    }
+}
+
+
+module.exports.Blockchain = Blockchain;
+module.exports.Transaction = Transaction; 
+
